Isolate legacy code paths in index.ts; drop WP 5.0–6.1 db.php placeholder
Extract every legacy mu-plugin and SQLite preload setup from
index.ts into dedicated files under legacy-wp/:

  * legacy-mu-plugins.ts — setupLegacyPlatformLevelMuPlugins,
    LEGACY_AUTO_LOGIN_BODY, legacy preload env.php, PHP 5.2
    error handler, legacy auto_prepend_file.
  * legacy-sqlite-preload.ts — preloadLegacySqliteIntegration,
    buildLegacySqlitePreload, mysql*/mysqli* stubs, str_*
    polyfills, WP < 3.1 add_action() guard.

Shared helpers live in neutral files at src/ root so legacy-wp/
doesn't import from the parent index.ts:

  * platform-mu-plugins.ts — writeCommonPlatformMuPlugins
    (0-playground.php, sitemap-redirect.php,
    inline-tinymce-content-css.php), called from both paths.
  * sqlite-preload-loader.ts — SQLITE_PRELOAD_LOADER_CLASS,
    referenced by both modern and legacy preload builders.

setupPlatformLevelMuPlugins and preloadSqliteIntegration now
dispatch to the legacy helpers at the top and their bodies carry
no phpMajor / isPhp52 branches. index.ts goes from 1521 lines to
691.

Also drop the WP 5.0–6.1 db.php placeholder.
backportWpPreV62MysqlCheck used to write an empty
wp-content/db.php so install.php step=2's \$mysql_compat check
would fall through its file_exists(WP_CONTENT_DIR . '/db.php')
escape hatch. The placeholder was never needed —
WP_SQLite_DB::db_version() returns '8.0' (per the SQLite plugin's
docstring: "it returns mysql version number, but it means nothing
for SQLite. So it returns the newest mysql version"), which
satisfies install.php's version_compare check directly via the
lazy-\$wpdb loader. With this change no file is ever written to
wp-content on the modern boot path.
